Change double to decimal.

+++ Tribby.Core/Models/Group.cs
@@ -5,8 +5,7 @@ public class Group
 
     public int Id { get; set; }
 
-    public double Balance { get; set; }
+    public decimal Balance { get; set; }
 
     public List<User> Users { get; } = new ();
-
 }
